Remove view logic from route
